+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - refactored some utility functions in helpers/index.js
   - minor typo fixes
   - formatting changes
+  - add some error checking for inputs and filetypes
 - Frontend
   - enabled .xlsx export (adds button and disables it for reports with no xls template)
   - added some checks to ensure proper options are passed to cdogs api